pyproject.toml
[tool.poetry]
name = "erddap-deploy"
version = "1.2.0"
description = "Run a series of tests on the different ERDDAP components which helps resolving the different issues related to an ERDDAP deployment"
authors = ["Jessy Barrette <30420025+JessyBarrette@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "erddap_deploy"}]

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.4.3"
lxml = "^4.9.3"
xarray = "^2023.12.0"
click = "^8.1.7"
loguru = "^0.7.2"
gitpython = "^3.1.41"
uptime-kuma-api = "^1.2.1"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^23.12.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
erddap_deploy = 'erddap_deploy.cli:cli'